Fix warnings about unbound record labels in patterns and unused variables
I also added a newline to fix a warning about an ambiguous documentation
comment, which was intended for the map function in the List module.

There are also warnings about unused names in the List module, which are
intended to shadow names from the standard library. I've disabled those
warnings for now.